Do it properly
